requirements/aws.txt
-r base.txt
boto3==1.17.36
django-sslify==0.2.7
django-ses==3.5.0
uWSGI==2.0.17.1
django-storages